refactor: extract dialog fragments from Anki[Activity/Fragment]
* dismissAllDialogFragments
* showDialogFragment

**AnkiFragment**
* showDialogFragment

This affects all callers, as we now use extension methods
