+++ AnkiDroid/src/main/java/com/ichi2/anki/AnkiActivity.kt
@@ -42,9 +42,7 @@ import androidx.browser.customtabs.CustomTabsIntent.COLOR_SCHEME_SYSTEM
 import androidx.core.app.NotificationCompat
 import androidx.core.app.PendingIntentCompat
 import androidx.core.content.ContextCompat
-import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
 import com.google.android.material.color.MaterialColors
 import com.google.android.material.snackbar.Snackbar
 import com.ichi2.anim.ActivityTransitionAnimation
@@ -61,11 +59,11 @@ import com.ichi2.anki.dialogs.DatabaseErrorDialog
 import com.ichi2.anki.dialogs.DatabaseErrorDialog.DatabaseErrorDialogType
 import com.ichi2.anki.dialogs.DialogHandler
 import com.ichi2.anki.dialogs.SimpleMessageDialog
-import com.ichi2.anki.dialogs.SimpleMessageDialog.SimpleMessageDialogListener
 import com.ichi2.anki.preferences.PENDING_NOTIFICATIONS_ONLY
 import com.ichi2.anki.preferences.sharedPrefs
 import com.ichi2.anki.receiver.SdCardReceiver
 import com.ichi2.anki.snackbar.showSnackbar
+import com.ichi2.anki.utils.ext.showDialogFragment
 import com.ichi2.anki.workarounds.AppLoadedFromBackupWorkaround.showedActivityFailedScreen
 import com.ichi2.async.CollectionLoader
 import com.ichi2.compat.CompatHelper.Companion.registerReceiverCompat
@@ -81,7 +79,7 @@ import androidx.browser.customtabs.CustomTabsIntent.Builder as CustomTabsIntentB
 
 @UiThread
 @KotlinCleanup("set activityName")
-open class AnkiActivity : AppCompatActivity, SimpleMessageDialogListener, ShortcutGroupProvider, AnkiActivityProvider {
+open class AnkiActivity : AppCompatActivity, ShortcutGroupProvider, AnkiActivityProvider {
 
     /**
      * Receiver that informs us when a broadcast listen in [broadcastsActions] is received.
@@ -456,18 +454,6 @@ open class AnkiActivity : AppCompatActivity, SimpleMessageDialogListener, Shortc
         }
 
     /**
-     * Global method to show dialog fragment including adding it to back stack Note: DO NOT call this from an async
-     * task! If you need to show a dialog from an async task, use showAsyncDialogFragment()
-     *
-     * @param newFragment  the DialogFragment you want to show
-     */
-    open fun showDialogFragment(newFragment: DialogFragment) {
-        runOnUiThread {
-            showDialogFragment(this, newFragment)
-        }
-    }
-
-    /**
      * Calls [.showAsyncDialogFragment] internally, using the channel
      * [Channel.GENERAL]
      *
@@ -569,27 +555,6 @@ open class AnkiActivity : AppCompatActivity, SimpleMessageDialogListener, Shortc
             val notificationManager = getSystemService(NOTIFICATION_SERVICE) as NotificationManager
             // mId allows you to update the notification later on.
             notificationManager.notify(SIMPLE_NOTIFICATION_ID, builder.build())
-        }
-    }
-
-    // Handle closing simple message dialog
-    override fun dismissSimpleMessageDialog(reload: Boolean) {
-        dismissAllDialogFragments()
-        if (reload) {
-            val deckPicker = Intent(this, DeckPicker::class.java)
-            deckPicker.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_NEW_TASK)
-            startActivity(deckPicker)
-        }
-    }
-
-    // Dismiss whatever dialog is showing
-    fun dismissAllDialogFragments() {
-        // trying to pop fragment manager back state crashes if state already saved
-        if (!supportFragmentManager.isStateSaved) {
-            supportFragmentManager.popBackStack(
-                DIALOG_FRAGMENT_TAG,
-                FragmentManager.POP_BACK_STACK_INCLUSIVE
-            )
         }
     }
 
@@ -721,29 +686,9 @@ open class AnkiActivity : AppCompatActivity, SimpleMessageDialogListener, Shortc
         get(): ShortcutGroup? = null
 
     companion object {
-        const val DIALOG_FRAGMENT_TAG = "dialog"
 
         /** Extra key to set the finish animation of an activity  */
         const val FINISH_ANIMATION_EXTRA = "finishAnimation"
-
-        fun showDialogFragment(activity: AnkiActivity, newFragment: DialogFragment) {
-            showDialogFragment(activity.supportFragmentManager, newFragment)
-        }
-
-        fun showDialogFragment(manager: FragmentManager, newFragment: DialogFragment) {
-            // DialogFragment.show() will take care of adding the fragment
-            // in a transaction. We also want to remove any currently showing
-            // dialog, so make our own transaction and take care of that here.
-            val ft = manager.beginTransaction()
-            val prev = manager.findFragmentByTag(DIALOG_FRAGMENT_TAG)
-            if (prev != null) {
-                ft.remove(prev)
-            }
-            // save transaction to the back stack
-            ft.addToBackStack(DIALOG_FRAGMENT_TAG)
-            newFragment.show(ft, DIALOG_FRAGMENT_TAG)
-            manager.executePendingTransactions()
-        }
 
         private const val SIMPLE_NOTIFICATION_ID = 0
     }
