+++ AnkiDroid/src/main/java/com/ichi2/anki/dialogs/ExportReadyDialog.kt
@@ -25,6 +25,7 @@ import com.ichi2.anki.AnkiActivity
 import com.ichi2.anki.DeckPicker
 import com.ichi2.anki.R
 import com.ichi2.anki.showError
+import com.ichi2.anki.utils.ext.showDialogFragment
 import com.ichi2.utils.negativeButton
 import com.ichi2.utils.positiveButton
