+++ AnkiDroid/src/main/java/com/ichi2/anki/instantnoteeditor/EditorChips.kt
@@ -28,6 +28,7 @@ import com.google.android.material.chip.ChipGroup
 import com.ichi2.anki.AnkiActivity
 import com.ichi2.anki.R
 import com.ichi2.anki.dialogs.IntegerDialog
+import com.ichi2.anki.utils.ext.showDialogFragment
 import com.ichi2.annotations.NeedsTest
 import timber.log.Timber
 
@@ -134,7 +135,7 @@ private fun AnkiActivity.displayUpdateClozeNumberDialog(clozeNumber: Int, newClo
             newClozeNumber(it)
         }
     }
-    this.showDialogFragment(repositionDialog)
+    showDialogFragment(repositionDialog)
 }
 
 private fun updateChipAndBadge(data: ChipData) {
